:art: [#296] -- restructure authentication module tests
